+++ api-derive/staking/erasHistoric.ts
@@ -2,14 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Observable } from 'https://esm.sh/rxjs@7.5.7';
-import type { u32 } from 'https://deno.land/x/polkadot@0.2.15/types/mod.ts';
-import type { EraIndex } from 'https://deno.land/x/polkadot@0.2.15/types/interfaces/index.ts';
-import type { BN } from 'https://deno.land/x/polkadot@0.2.15/util/mod.ts';
+import type { u32 } from 'https://deno.land/x/polkadot/types/mod.ts';
+import type { EraIndex } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
+import type { BN } from 'https://deno.land/x/polkadot/util/mod.ts';
 import type { DeriveApi } from '../types.ts';
 
 import { combineLatest, map, of } from 'https://esm.sh/rxjs@7.5.7';
 
-import { BN_ONE, BN_ZERO } from 'https://deno.land/x/polkadot@0.2.15/util/mod.ts';
+import { BN_ONE, BN_ZERO } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { memo } from '../util/index.ts';