+++ api/types/consts.ts
@@ -1,4 +1,4 @@
 // Copyright 2017-2022 @polkadot/api authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-export * from 'https://deno.land/x/polkadot@0.2.15/api-base/types/consts.ts';
+export * from 'https://deno.land/x/polkadot/api-base/types/consts.ts';