+++ types/interfaces/payment/definitions.ts
@@ -27,6 +27,16 @@ export default {
       weight: 'Weight',
       class: 'DispatchClass',
       partialFee: 'Balance'
+    },
+    RuntimeDispatchInfoV1: {
+      weight: 'WeightV1',
+      class: 'DispatchClass',
+      partialFee: 'Balance'
+    },
+    RuntimeDispatchInfoV2: {
+      weight: 'WeightV2',
+      class: 'DispatchClass',
+      partialFee: 'Balance'
     }
   }
 } as Definitions;