+++ types/interfaces/payment/types.ts
@@ -1,9 +1,9 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable */
 
-import type { Option, Struct } from 'https://deno.land/x/polkadot@0.2.15/types-codec/mod.ts';
-import type { Balance, Weight } from 'https://deno.land/x/polkadot@0.2.15/types/interfaces/runtime/index.ts';
-import type { DispatchClass } from 'https://deno.land/x/polkadot@0.2.15/types/interfaces/system/index.ts';
+import type { Option, Struct } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
+import type { Balance, Weight, WeightV1, WeightV2 } from 'https://deno.land/x/polkadot/types/interfaces/runtime/index.ts';
+import type { DispatchClass } from 'https://deno.land/x/polkadot/types/interfaces/system/index.ts';
 
 /** @name FeeDetails */
 export interface FeeDetails extends Struct {
@@ -20,6 +20,20 @@ export interface InclusionFee extends Struct {
 /** @name RuntimeDispatchInfo */
 export interface RuntimeDispatchInfo extends Struct {
   readonly weight: Weight;
+  readonly class: DispatchClass;
+  readonly partialFee: Balance;
+}
+
+/** @name RuntimeDispatchInfoV1 */
+export interface RuntimeDispatchInfoV1 extends Struct {
+  readonly weight: WeightV1;
+  readonly class: DispatchClass;
+  readonly partialFee: Balance;
+}
+
+/** @name RuntimeDispatchInfoV2 */
+export interface RuntimeDispatchInfoV2 extends Struct {
+  readonly weight: WeightV2;
   readonly class: DispatchClass;
   readonly partialFee: Balance;
 }